Github_Webhook_Listener_and_Topic_Emitter/requirements.txt
fastapi>=0.65.0
uvicorn>=0.14.0
pydantic>=1.8.2
confluent-kafka>=1.7.0
requests>=2.25.1
tenacity>=8.0.1
PyJWT>=2.1.0
Jinja2>=3.0.1
python-dotenv>=0.18.0
pydantic-settings
python-multipart
kafka-python-ng
PyGithub
icecream